scripts/topics_creator/requirements.txt
ConfigArgParse~=1.5.3
kafka-python~=2.0.2
PyYAML~=6.0